Fix split package in guice module
`org.elasticsearch.common` is used by the server module.
